Add epa_build configuration for libLASi.
Tested on Linux using the build_libLASi target.  That target
configures, builds, installs, and tests (via ctest) libLASi. The ctest
results look fine except that the detailed PostScript file results for
the second example are different each time that example is run as
discussed on the lasi-devel list.  It appears the font commands at the
head of the PostScript file are created in different order with
different names, but the results look visually identical (as expected
if just the font order is changed each time).  So this bug is more a
curiosity than not, but nevertheless it would be good to deal with it
so that the second example produces file-identical results each time
rather than different random orders of fonts.


svn path=/trunk/; revision=13131