Switch to a different OCaml Cairo binding ("cairo2" in opam)
This is a more actively maintained binding to the Cairo library.  It is
unfortunately not packaged for Debian at this time so interested users
would need to install the library through opam.  opam is generally the
recommended way to install and manage OCaml libraries (this cpan, pip, etc)
so this isn't a particularly serious limitation.

svn path=/trunk/; revision=13134