Preliminary update of documentation of release procedure due to replacing svn with git.
The file edits included tracking down every instance of "svn" and
"trunk" and making appropriate changes for the git case.  Also, I
took this opportunity to document the new options for rsync that
give a clean upload of the website with (I hope) no known
permissions problems for the next time someone else tries
to generate and rsync the website.