Store information concerning our repository conversions.
The information about our historical conversion of the PLplot CVS repo
to svn is summarized from e-mail messages at the time. The information
about our recent conversion of our svn repo to git has all the details
of the conversion including all files used in the conversion and in
testing that converted result.