Only do try_compile concerning HAVE_SAHOOKS if the information is not cached already.
This is supposed to happen in any case according to CMake
documentation, but does not.  So for the --debug-trycompile case
(which explicitly does not clean up CMakeFiles/CMakeTmp after each
try_compile) cannot follow documentation suggestion to simply remove
from the cache the try_compile result you want to debug.  Therefore,
have to put in an explicit if statement checking the existence of the
cache value for all try_compiles to make that suggestion work.