Change line ending handling for *.sh, *.sh.in, *.txt, and *.cmake files
*.sh and *.sh.in were changed to eof=lf to force LF endings in all
cases for those files, and *.txt and *.cmake were
changed to text=auto to for LF in repository and native on checkout.
Note, this last change caught some LF errors in two files.  The
plplot_test/CMakeLists.txt file was a known issue with a recent commit,
but the CMakeDInformation one was a complete surprise.  Which
shows how easy it is for line endings issues to creep in.