Treat most text file line endings as LF in repository and native in working directory
This change means that almost all of our text files will have LF
endings in the repository and native (LF on Unix and CRLF on Windows)
line endings in the working directory.

The text file exceptions are *.sh and *.sh.in which are treated as LF
both in repository and in the working directory so that these bash
shell scripts will work on both Unix and Windows.

Also, just in case the git heuristic for classifying whether a file is
binary or text is making the wrong classification of a file as text
when it is actually binary, specifically classify a number of file
patterns as binary.  These file patterns were determined by looking at
"svn proplist --verbose --recursive" results for the entire source
tree of our last svn revision for svn trunk and searching for
"application/octet-stream" in the results.

Note this change in .gitattributes (when combined with the rest of the
git commands that are recommended for each .gitattributes change by
the gitattributes man page) found four additional text files where
CRLF line endings had crept into the repository version. This commit
fixes those issues as well.