Modify recent commits concerning Tcl testing on Windows
These changes were the following:

(a) The Cygwin, MSYS, and MinGW build-system path logic was left as it
was before Phil's recent changes (except for Phil's blank in path
changes) since that path logic was previously extensively tested
on the given platforms.

(b) Change to a style where the platform is determined by CMake
as opposed to running uname at the bash level to determine the
platform.

(c) An attempt was made to keep all of Phil's changes for the MSVC
case in the configured test_tcl.sh file.  There were some exceptions however.

I. native_windows=true was dropped.  I could not find any reference to
that with google so I suspect setting that in bash does nothing.  But
if it does anything, it is possible this is the source of the
difficulty with the drive letter logic experienced by Phil on Cygwin.

II. PLTCL_DIR_NATIVE=$PLTCL_DIR was dropped since apparently that
variable is not used.

III. Dropped the extra cd "$EXAMPLES_DIR/tcl" commands.  Instead
the script uniformly uses cd "$tcldir" which is consistent with
the design of the script before Phil's changes.  If something
special has to be done for the MSVC platform case, then
tcldir should be set appropriately in plplot-test.sh.in depending
on the value of PLATFORM=@PLATFORM@ in similar style to the
current changes in plplot_test/test_tcl.sh.in.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
the test_tcl target in the build tree where the build-tree path had no
blank in the name.  Note, the test_tcl target currently fails if there
is a blank in the build-tree name, but I presume that is because Phil
is not done yet with sorting out the "blank-in-build-tree-name"
issues.