Rename some source files for the f95 bindings
The new names just make more sense to me.  "configurable.f90" wasn't
configurable so I renamed that plparseopts.f90 for the name of the one
subroutine in it and "sfstubs.f90" was a fortran include file so I
renamed that to sfstubs.inc to remind of that important distinction.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
the test_f95_psc target.