doc: Remove extremely dated file describing PLplot
Historically, abs.txt was a way to save unconverted parts of the
texinfo documentation of PLplot that was transformed into the modern
DocBook form of documentation.  But by 2004 this file had degenerated
into just a description of PLplot.  This file has not been changed
since, that description is now quite incomplete, and we have different
ways (e.g., the README file in the top-level deirectory) to describe
PLplot in any case.  So remove this file.