doc: License man files under the FreeBSD documentation license
This license is identical to the one used for our DocBook src files.
Rafael Laboissiere has recently agreed to this explicit licensing for
the man files (which he wrote a decade ago).

Note the licensing text is embedded in comments within the man page
files.  I believe that should be sufficient, and I have therefore
decided not to clutter the output of the man command with the
licensing terms for those man pages.

Note that emacs needs no special directive to put itself in nroff mode
for these man page files.  So as part of these changes I have removed
the first comment line with that special emacs directive in all cases.