Python: Propagate new C plmap* API to Python binding and Python example 19
Note the implementation was quite straightforward because I knew the
SWIG typemap that had to be defined for the Python case because of
the prior experience with Lua, and because the Python syntax used for
the additional example 19 page is not that different from the Lua
case.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using test_diff_psc target.  The result was no example 19
differences for Python.