C, C++: update example 8.
The changes are as follows:

The sombrero function is now the default with the log_e of the
Rosenbrock function the option.

The index ranges for pages 5 and 10 are now elliptical and symmetrical
around the centre of the sombrero.

The semimajor and semiminor axes of the ellipse and the altitude and
azimuth of the viewing angle have been adjusted to give nice-looking
plots for pages 5 and 10.

These changes should be finalized now so that either the C or C++
version can be used as a template for propagation to other
languages.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using the x08c and x08 targets and running

examples/c/x08c -dev psc -o test1.psc
examples/c++/x08 -dev psc -o test2.psc
diff test[12].psc
examples/c/x08c -rosen -dev psc -o test1.psc
examples/c++/x08 -rosen -dev psc -o test2.psc
diff test[12].psc

To show the C and C++ results were identical (except for date)
for the default (sombrero) case as well as the rosen case.