C++: bug fix for fill3 and gradient for the multistream case.
Previously, fill3 and gradient were incorrectly declared and defined
following what was done for the fill special case.  The static
attribute was used for the declaration and the set_stream call was
commented out inside their definitions.  This will potentially lead to
trouble for direct use of fill3 and gradient for the multistream case
so I have removed the static attribute and called set_stream for these
functions to make them conform to virtually all other C++ API.

Note that fill3 and gradient are never used as callbacks which is why
the above fix can be made without breaking API.  A similar fix is
needed for the fill case, but the complication there is fill is used
both directly (where static attribute should not be used and
set_stream should be called) and also as a callback where the static
attribute and no set_stream call should be used. so C++ API breakage will
be required to distinguish those two cases.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using the test_diff_psc target.