ps and psttf device drivers: fix reinitialization of current point location.
The current change fixes a two-decade old fix in the ps device driver
for the M (moveto) command after the C
(setrgbcolor) command which prior to that old fix depended on
coordinates which were sometimes undefined.  However, that old fix was
implemented in a way that affected every change in state, not just the
particular state calls for PLSTATE_COLOR0 and PLSTATE_COLOR1 that
emitted C commands, see
<http://sourceforge.net/p/plplot/plplot/ci/07eab71b70249ac03a0f89db4061583d968f7365>

I have now changed the fix so that it only occurs for the relevant
PLSTATE_COLOR0 and PLSTATE_COLOR1 cases.  The result is that the ps
device driver is now much less sensitive to rounding issues, which do
cause additional state changes which now do nothing but which before
this fix triggered bare M commands.  For example, the results for the
(as yet uncommitted) changes for Python and C example 8 disagreed
before this fix with Python emitting two extra M commands, but those
Python and C example 8 results now agree as a result of this fix.

The old fix was propagated to the psttf device driver code so
I fixed that as well.  Before the current fix, Python and C
results for the psttfc device differed for example 8 with the
Python version emitting two extra M commands, but now the
psttfc results for those two languages agree.