Fix file dependency issue for the docbook_plplot-symbols_txt custom_target.
From Phil's recent evidence it appears that a
docbook_plplot-symbols_txt target dependency is not enough to assure
the plplot is always built before it is needed by the custom command
associated with the docbook_plplot-symbols_txt target, and you need a
file dependency as well.

Follow the method suggested by the CMake documentation of specifying
the plplot target in the DEPENDS of the associated custom command to
automatically take care of both file and target dependencies on the
plplot target.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
the docbook_plplot-symbols_txt target after a "make clean".  As
expected from the CMake documentation, the plplot target was run
before the docbook_plplot-symbols_txt target was run.  Furthermore, a
second run of the docbook_plplot-symbols_txt target produced no
further changes (again as expected from the CMake documentation).