Document additional software we need for the Documentation build.
I updated the list of applications that are required
to do validation and all parts of the documentation build.
I carefully went through cmake/modules/plplot.cmake and
cmake/modules/docbook.cmake to find everything required
so I am pretty sure this list is now complete.

I also moved the "Cygwin Specific" section to a more logical location
and wrote a similar "Linux Specific" section that lists exactly what
packages need to be installed to build the documentation on Debian and
gives what I hope is useful advice for building the documentation on
rpm-based Linux distributions.