Configure the installed PLplot header location in plplotqt.pro
According to
<http://qt-project.org/doc/qt-4.8/qmake-project-files.html> qmake
project files such as plplotqt.pro are a similar concept to pkg-config
files; they summarize everything relevant to building a project in a
nice compact format (in this case for qmake rather than pkg-config).

According to
<http://qt-project.org/doc/qt-4.8/qmake-variable-reference.html#includepath>
INCLUDEPATH is an important variable that is required to find the
appropriate headers.  So simply removing this variable is not a good
option since it will screw up any qmake user that is using anything
but the "/usr" install prefix for PLplot.  So I have instead made
plplotqt.pro configurable with INCLUDEPATH specified in the same
configurable way as is done in pkgcfg/plplot-template.pc.in when
configuring the header location for the pkg-config case.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by running cmake and looking at the configured
bindings/qt_gui/smoke/plplotqt.pro file in the build tree;
the configured header location is correct in that file.

ToDo: Figure out where the configured plplotqt.pro file
should be installed, and then use the appropriate CMake command
to install it there.