Use generator expression to find location of library used to get symbols for documentation needs.
Previously this location was specified as src/libplplot.so in the build
tree which worked fine on Linux, but a generator expression is a more general cross-platform
way to specify locations, and may solve the "double build" workaround currently
necessary to build documentation on Cygwin.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
the CMake option -DBUILD_DOCon and the "pdf_target" target (which
builds the pdf form of our documentation).