Build System: Fix bug for case where no display
AWI for Arjen Markus.

The issue was one of the Tcl commands used by our build system to
find Tcl packages.  The Tcl command "package require" does not work if
there is no display so replace it with "package provide" which
does work for this situation and also the case when there is a display.