Build system.  Revert recent change to the Tk configuration
It turns out that recent change (to use "package provide" rather than
"package require" for one CMake test to find the Tk version in
tcl-related.cmake) never works on Linux.  So that change was a
roundabout way to insure that Tk was always disabled which is not the
desired result.

I have done extensive command-line tests, and "package provide" only
appears to work with Tcl (presumably because tclsh loaded Tcl before
that package command was executed), but not Itcl, Tk, Itk, or Iwidgets
under either tclsh or wish.  So the conclusion is we must always stick
with "package require" in tcl-related.cmake and accept that a working
X will be required to configure Tk, Itk, or Iwidgets.  The logic in
tcl-related.cmake does look for the case where either wish or tclsh
error out.  However, if X is hanging rather than erroring out, (the
issue which was "solved" by the commit that this commit reverts), I
don't know a test for that issue that would not also hang.  Instead,
it appears to me the only recourse for a user if X hangs is to specify
-DENABLE_tk=OFF (which automatically turns off searching for or using
Itk and Iwidgets as well).