Added obsolete commands to plbuf_control to attempt some backwards compatabilty with old plot metafiles.
Removed unneeded malloc() and memcpy for LINE, POLYINE, and FILL when
remaking the plot.  Temporary storage was used because of the old
file-based buffer approach.  With the removal of file-based buffer
code, a different apporach that uses pointers into the plot buffer could
be used.

Fixed the bug that caused random segfaults.  This could be manifested
in several ways, one of them being a POLYLINE with over a billion nodes.
The root cause was a mismatch in the data sizes between the buffer writer
and the buffer reader (a wrong sizeof() was used).  Changed the code
to use sizeof(variable) instead of sizeof(data type).

Changed the buffer writer to maintain two-byte alignment.  This should
avoid the performance penalty associated with misaligned arrays.