libplplot: Fix buildability (visibility) issue with plreadmetafile
I fixed this visibility issue by declaring this function in plplotP.h with
the appropriate PLDLLIMPEXP macro.  If we want more visibility for
this function, the declaration should be moved to include/plplot.h, but
we start with this declaration so that nobody experiences build issues
for the shared library case on Linux (with
-fvisibility=hidden gcc option) and presumably on Windows.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by building the test_plbuf target using -fvisibility=hidden.