Style fixup
This style update was generated by the

scripts/style.sh --apply

command (which in turn runs a python script
to put comments in c99 "//" form and uncrustify
to deal with all other style issues).

This time, it was a large and intrusive change because I had
waited for too long to run the script, and also because so
much plbuf and wxwidgets work has landed on master recently.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by running the test_noninteractive target in the build tree as
well as running several standard examples with the wxwidgets and xwin
devices and resizing those results.  I also looked at the results
from

git diff --ignore-all-space |less

which didn't show any obvious problems in the restyling but did
graphically illustrate (because of the change of comments to "//"
style) that large parts of drivers/wxwidgets.cpp and
examples/c++/wxPLplotDemo.cpp had been commented out.