Fixes to reading plot metafiles
The original plan to utilize the plot buffer did not work because the
coordinate system used in the plot metafile is not compatible with
the device level coordinate system that is in the plot buffer.

In order to keep redraws (one of the key uses of the plot buffer) fast,
the same strategy used by plrender was adopted.  The metafile coordinate
system is used to setup a coordinate system via plvpor() and plwind().

This will need to be revisited for the next revision to the plot buffers
that will support switching of devices to handle saving/printing.