plrender: replace call to plSetOpt with call to plsetopt
plSetOpt calls plsetopt with an identical API, but the plSetOpt spelling
is deprecated and requires use of -DPL_DEPRECATED=ON to work properly
so use the recommended plsetopt instead.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by setting the cmake option -DPLD_plmeta=ON and building the plrender target.