libplplot: fix issues with long-standing plSetOpt deprecation
plSetOpt has been deprecated a long time, but the original
implementation of that deprecation had some issues which are
now fixed.  Those issues were the following:

* Internal use of plSetOpt. This has now been completely fixed;
plsetopt has replaced plSetOpt in a previous plrender fix, and for
this commit, an unused plSetOpt declaration was removed in the D
binding.

* src/plargs.c declared plSetOpt for the case where
-DPL_DEPRECATED=OFF to compensate for the lack of declaration
in include/plplot.h for that case, but now that there is no
internal use of plSetOpt, there is no need for that workaround
and instead plSetOpt is now only declared
and defined when -DPL_DEPRECATED=ON.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using the test_noninteractive and install targets in the build tree.