Add missing routines to the definition file for gfortran under MinGW/MSYS.
Also changed the name of the library in accordance with the name that is actually used in the link step. Quite possibly the first line needs to be removed altogether. This requires some testing though.

Note: I (Arjen Markus) could not test it because the current version of gfortran
under MinGW does not seem to need such a file. The additional routines were taken from the corresponding definitions file for Intel Fortran on Windows.