Save device interface information to the buffer, this allows redraws performed by the buffer to correctly utilise orientation changes or other transforms set via the pldi_ini(0 function and functions that call this function.
Modified the wxWidgets driver to correctly take advantage of these
changes which required setting the PLStream aspect function and some
rewriting of the sizing code which allows fixed aspect rendering.

Modified the wxPLplotwindow class to fully clear the background on a
redraw so that this reqirement is removed from the driver allowing it to
just clear the subpages.