Build systems: Fix configured test_tcl.sh for build tree and installed examples tree
* Fix value of PLTCL_DIR_ESCAPED used to configure installed version
of test_tcl.sh.  This fixed a bug in the test_noninteractive results
for the traditional version of the installed examples tree where
core build-tree results were referenced rather than the expected
install-tree results.

* For build-tree and install-tree versions of test_tcl.sh remove many
references to $tcldir/ since that can be a relative path location that
fails when the script has already changed directory to $tcldir.  (This
issue discovered when running plplot-test.sh on its own in the
traditionally built version of the installed examples tree.)

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by building test_noninteractive in the core build tree and
for the traditional version of the installed examples tree and
by running plplot-test.sh on its own in the
traditionally built version of the installed examples tree.