libplplotwxwidgets: add target_link_library dependency on plplot
This is necessary because
this library directly uses a call to the function plsstrm for
setting a stream number that is defined in the plplot library rather
than plplotcxx library.  I don't understand this direct dependence,
but nm confirms this as follows:

nm --undefined-only --demangle bindings/wxwidgets/libplplotwxwidgets.so |grep plsstrm
                 U c_plsstrm

So target_link_library necessarily must
mention plplot in order for linking to work properly on all platforms.

This investigation and fix was motivated by a report from Jim Dishaw
that indicated the the linking of libplplotwxwidgets was failing
on Mac OS X with a reference to _c_plsstrm (which presumably is the
Mac OS X linker way to mangle c_plsstrm).

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by building the test_c_wxwidgets target in the build tree.