Added plflush capability to the wxWidgets driver and changed the wxPLDevice::ClearBackground function so that it draws a rectangle using plplot functions. The latter change was required because when replotting from a buffer we have no concept of subpages and the subpage size is not stored in the buffer meaning we have no way to know how big an area to erase.
Also added calling plP_bop() to rdbuf_bop. This is required to do
the initial page setup. Previously in the wxWidgets driver this was
performed by a manual call to wxPLDevice::BeginPage() from within
wxPLDevice::SetSize() before calling plreplot(). The previous
solution invalidated the principle of least surprise, but then also
meant that the initial ClearScreen call appended drawing the above
rectangle to the end of the buffer and plreplot then ended by erasing
everything.

This makes the Strip chart demo x17 work for wxWidgets now,
although it is rather slow due to inefficient buffer copies.