wxPLViewer: drop explicit link to the rt library
That explicit link causes trouble on Mac OS X (which does not
support the rt library even though that is a POSIX standard)
and is not required on Debian stable.  Testing of additional
Linux systems like Ubuntu and CentOS may show rt has to
be explicitly linked in one/both of those cases in which
case the fix is to uncomment one statement.