"definitive fix" version 2 for the rt linking issues
I must have made a grep mistake with my nm analysis before because it
turns out that both the wxwidgets device driver and wxPLViewer source
code includes drivers wxwidgets_comms.cpp which contains the key calls
to shm_open and shm_unlink, and the corrected nm analysis supports
that conclusion.

For some reason Ubuntu worked with the previous explicit linking of
wxPLViewer with librt but no explicit linking of the wxwidgets dll
with that library.  This may be due to differences between dynamic
loading of the wxwidgets dll and run-time loading of the wxPLViewer
application. But best practice requires that dll and utility both to
be explicitly linked to the rt library for all Unices other than Mac
OS X so that is what is done for this commit (with the necessary move
of the required determination of the CMake RT_LIB variable from
utils/CMakelists.txt to the cmake/modules/wxwidgets.cmake location
which makes RT_LIB universally accessible to all CMake code).

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using the test_c_wxwidgets in the build tree.  Also, before the
fix the nm analysis gave this result

software@raven> nm -a --demangle utils/wxPLViewer |grep shm_
                 U shm_open@@GLIBC_2.2.5
                 U shm_unlink@@GLIBC_2.2.5
software@raven> nm -a --demangle drivers/wxwidgets.so |grep shm_
                 U shm_open
                 U shm_unlink

i.e., nm referred to undefined unversioned symbols for wxwidgets.  But
after the fix these undefined symbols became versioned just like the
symbols for wxPLViewer which I assume is an indicator that the
undefined symbols are resolved by the particular

-- RT_LIB = /usr/lib/x86_64-linux-gnu/librt.so

that is found on my system and which is used to explicitly link
both wxPLViewer and wxwidgets now.