wxwidgets device driver: Replaced undocumented wxFont method with documented variant
The official wxwidgets documentation does not document the wxFont
overloaded variant that we were using which was likely taken from a
much earlier version of wxwidgets (but still supported with
wxwidgets-2.8.12 for backwards compatibility).  So to follow best
practice and also to follow what is done with the font sample
(demonstration code showing how to display wxwidgets accessible fonts
in a GUI), I changed to using a documented variant of wxFont to set up
the font.  The hope was this change would sort out the missing glyph
issue, since the font sample GUI displays all the utf-8 Peace words
without issues.

However, the wxwidgets device results for all unicode examples are
exactly the same as before (two missing glyphs from example 23 and
numerous missing glyphs frome example 24).  So the missing glyph
issue for wxwidgets is likely due to something else then the
font setup.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux by
building the wxwidget device (and all the C examples) in the build
tree then running the unicode (18th, 23rd, 24th, and 26th) examples by
hand with -dev wxwidgets and looking for missing glyphs in the
results.

This change was suggested by the font sample