pkg-config: Convert Qt5::<component> imported library names to a form that pkg-config can understand
The Qt5 cmake logic that is part of Qt development (as opposed to
CMake development) uses the name convention that Qt5::<Module> is the
name of the imported library target name for each Module.  This change
translates those target names to file locations that pkg-config can
understand.  Part of this change is to translate those target names to
disk locations which typically include a trailing version number, e.g.,
<some prefix>/libQt5Core.so.5.3.1.  So the current change also required
modifying the suffix_pattern used in the pkg_config_link_flags
function for the .so case from \\.so to \\.so.* to allow trimming off
that extra version information when reducing absolute disk locations to
the standard pkg-config form such as

-L<some prefix> -lQt5Core

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux for an
epa_built Qt5 and Plplot case.  The comprehensive test for that case
was a success for the shared library/dynamic device configuration.
(Before this fix that comprehensive test configuration had failed for
the sub-case of the traditionally built installed examples tree
because pkg-config (obviously) cannot understand the untransformed raw
Qt5::<Module> CMake syntax.