Changed the wxwidgets driver and binding to make use of the plsdevdata function. Now we can det the wxDC using this function before the device initialisation allowing us to know at this point whether we want to draw on a user supplied wxDC or the separate wxPLViewer.
This was quite a major change and some other alterations were made at the same time.

The default dpi is now only used if the user has not called plspage. It then becomes the users
responsibility to set the dpi at the same time. If the user set the page size but not the dpi then
the driver issues a warning.

The wxPLplotwindow has had some major changes. We now require a dc available at all time for
drawing to rather than just in the OnPaint method. We had to make the dcs pointer created with
new. This was because changing the size of the bitmap being drawn to causes lots of problems (at
least on Windows) with wxGCDC unless we have a fresh wxMemoryDC, and (because there is no
wxGCDC::SetlectDC type method) this requires a new wxGCDC too.

Added try cath blocks to all the dispatch table functions to avoid throwing an exception into the
C  library.

Removed  all the attempts to guess whethe the user wanted to use their own dc.

made sure that during all drawing commands the pen and brush of the dc are reinstated to there
previous styles

changed from using wxThePenList to just creating pens as the former sometimes caused problems as
there seemed to be no pens in the list, resulting in assert fails

fixed bug where xdpi was used instead of ydpi in wxPLDevice::SetSize

Allowed wxPLViewerClearBackground to draw on the dc as beginPage will now actually be called
during a replot

Fixed bug in wxPLDevice::SetWidth where alpha channel was missed.