Fixed calls to plpat being missed from the buffer.
To do this I moved the plP_state call from c_plpsty to spat, which is called from both c_plsty and
c_plpat. Also both the pattern index and the pattern parameters (number of lines, inclination and
separation) are stored in the buffer. On replot the pattern style is set directly in the stream,
and the pattern parameters via a call to c_plpat, which ensures the driver is called as necessary.