Style previous commits
Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux. This
testing was done not only to test this commit, but the very large
series of wxwidgets/plbuf commits done recently.  I looked carefully
at the rendering of most pages for all examples for -dev wxwidgets
with the remaining issues reported to list (and eventually to
the wxwidgets overall bug report).

I also checked -dev tk for the zoom regression, and confirm that
regression as well as a newly introduced regression for that device
(two windows are displayed rather than one with one of them blank
with the plframe decorations and the other showing an xwin-like plot
windows with no decorations).
This new regression (and likely the zoom regression) are showstoppers
for the current release.

I did no testing of the new plmeta/plrender, and I also did no testing
to look for other regressions in rendering for non-wxwidgets devices.