epa_build: qt5_lite: Update Qt5 build configuration
Since making a complete build of Qt5 did not make any difference
to the invalid read memory management issue for qt_example and -dev
qtwidget, change the build back to a minimal build that is even
smaller than we tried before with Qt5.3.2.  This changes the
epa_build time of Qt5 from 2 hours to 13.5 minutes, and does
not compromise PLplot's use of Qt5.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
an epa_build of PLplot (which necessarily triggers an epa_build of
Qt5) and running tests on that PLplot version. Those tests included
valgrind runs of example 1 for -dev svgqt, pdfqt, pngqt, and bmpqt
(all with no memory management errors), valgrind runs with -dev
qtwidgets (which showed the invalid read), and a build of the
test_noninteractive target.