Build system: add the custom target build_docbook_documentation
When BUILD_DOC is ON this custom target builds all components of
DocBook documentation and their prerequisites.  This is somewhat
faster than the "all" target that has previously been used for this
task since "all" builds everything while build_docbook_documentation
only builds the necessary subset of PLplot required to generate the
documentation.