Build system: For shared libraries case always set USINGDLL macro for C examples
This should make an important improvement for the MSVC case where
from include/pldll.h.in, apparently USINGDLL has to be #defined for
the import case, i.e., for building the C examples (as opposed to the
export case when building libraries where USINGDLL has to be #defined
for all compilers).

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by building all in the build tree.  This is only a light test
to make sure there are no errors in the CMake logic.  The real
test comes later for the MSVC case, and if that is a success, then
must follow up with similar changes for all examples for all compiled
languages.