Build system: Solve qt_example segfault for ENABLE_DYNDRIVERS=NO case
For the ENABLE_DYNDRIVERS=NO case the qt (and other) device driver
code is stored in libplplot, but qt normally depends on the code
in libplplotqt which in turn depends on libplplot.  So to break this
circular linking chain, the libplplotqt code is also stored in libplplot
with libplplot not linking to (or depending on) libplplotqt.  So
the net result is we have identical code stored in both libplplot and
libplplotqt, and if we manipulate visibility so that qt_example uses
one or the other, all is well.

However, previously, qt_example was linked to both plplot and plplotqt
for the ENABLE_DYNDRIVERS=NO case which is redundant.  More seriously,
and for reasons I don't understand this choice of linking to the same
code in two different libraries caused severe memory management issues
when the libraries were closed.  So the fix for this issue was simply
to drop the redundant link to plplotqt for qt_example.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
for the shared libaries/static devices (ENABLE_DYNDRIVERS=NO) case
in the build tree by building plplot, plplotqt, and qt_example, testing
each of those with "ldd -r", and running examples/c++/qt_example
under valgrind.