Build system: Fix long-standing -DUSINGDLL bug
-DUSINGDLL should only be used as a compile flag for the shared
library case.  This bug only affected the build of the qt_gui and
smoke bindings on Windows for the static library case.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by running cmake.  (A simple test that no syntax errors were
introduced.)  An actual build of libplplotqt for the static library
case on Windows is needed for definitive testing of this fix.