Remove "c_" prefix from plsdevdata.
The plsdevdata function is currently used only by the wxwidgets device driver and
associated binding.  This function might be used for other device
drivers later on whether written in c++ or c.  But this function is not
part of the common API and should not be propagated to any binding other than
c++.  Therefore remove the "c_" prefix which is an incorrect indicator that this
function is part of the common API that should be propagated to all bindings.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
by building the test_c_wxwidgets target in the build tree.