Update scripts/make_tarball.sh from svn to git
In addition to modifying this script to work with git, I updated the
messages generated by the script, and (most importantly) fixed a
long-standing bug in the script where the exported working directory
was used for "testing the tarball" as opposed to the correct unpacked
source tree from the tarball.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
following the "Create (a preliminary version of) the release tarball
and check the result for errors" part of
README.Release_Manager_Cookbook.