Comprehensive testing bug-fix for required PATH manipulations on Windows platforms
This script uses the following PATH manipulations on Windows
platforms.

1. Put the dll subdirectory of the build tree on the PATH for _just_
the build-tree parts of the testing.

2. Put the installed bin directory and drivers directory on the PATH
for _just_ the install-tree parts of the testing.

As discussed on list prior to the 5.11.0 release, these required
Windows PATH manipulations were only being done on MinGW and
MinGW/MSYS which lead to script errors on Cygwin (unless the PATH was
brute-forced prior to running the script in a way that mixed
build-tree and install-tree results together).

As far as I know, these PATH manipulations should be done on all
Windows platforms.  Therefore, this bug fix detects Windows platforms
in a much more general way now (by looking for the "exe" suffix for
the full pathname of the bash command), and this should solve
the Cygwin issue (i.e., the brute-force PATH manipulations should no
longer be required) and similar issues that would have occurred for MSVC
once that platform was comprehensively tested with this script.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using

scripts/comprehensive_test.sh \
--do_nondynamic no \
--do_static no \
--do_ctest no \
--do_test_interactive no \
--do_test_traditional_install_tree no

i.e, run the test_noninteractive target in the build tree and
the non-traditional installed examples tree for the shared library
case only.

Note, this test merely checks that the script still works on Linux,
but the definitive test of this bug fix will require running this
script on Cygwin and other Windows platforms.