Build system: Adopt a minimum CMake version of 3.0.2 everywhere
In order for our current build system to work, we have had to
explicitly set several CMake policies to old, but using such
deprecated features of CMake is not ideal so I plan to make
changes in our build system so that it works with new policies.

The problem is, however, that those planned changes use CMake
functionality that is only available for CMake version 3.  Therefore,
as a prerequisite for these planned changes, in this commit I have
updated the minimum allowed version of CMake to 3.0.2 everywhere in
our source tree that the cmake_minimum_required command is used.